refactor getCollection and addContact
