+++ .env
@@ -1,1 +1,1 @@
-URI_DB=mongodb+srv://goit25:567234@rest-api-cluster.ybkgd.mongodb.net/contacts?retryWrites=true&w=majority
+URI_DB=mongodb+srv://goit25:567234@rest-api-cluster.ybkgd.mongodb.net/db-contacts?retryWrites=true&w=majority
